Move certificate builder code from core to node utilities.
Address PR comments

Fixup dead reference
